+++ app/components/ui/chakra-ui/navigation/index.ts
@@ -1,1 +1,2 @@
 export * from './CustomBreadcrumb';
+export * from './CustomButtonNavGroup';
