+++ app/routes/dashboard+/builder+/design-attributes+/index.tsx
@@ -1,9 +1,12 @@
-import { Stack, Text } from '~/components';
+import { CustomButtonNavGroup, Stack, Text } from '~/components';
 
 export default function DesignAttributesIndexPage() {
+  const navs = [{ to: 'container', title: 'Container' }];
+
   return (
     <Stack width="full" paddingX={8} paddingY={5} textAlign="left">
       <Text fontSize="sm">Select a Design Attribute</Text>
+      <CustomButtonNavGroup navs={navs} />
     </Stack>
   );
 }
